Named pipe can be binded either in message or stream mode
Function bindPipe now takes PipeMode argument, which is a different type
from the PipeMode as introduced in low-level Win32 bindings. It allows
us to decide if the pipe is created either in message oriented, or
(byte) stream oriented regime. This way we defer the decision to a
higher abstraction level.