Handle issues with non-idempotent close/disconnect/flush
Win API functions CloseHandle, DisconnectNamedPipe, and FlushFileBuffers
fail when handle is already close with ERROR_INVALID_HANDLE. For this
reason we need to detect this situation to allow nested brackets in server
and client implementations.

To get rid of this issue we had to drop to more low-level plumbing that
previously. Now we have our own wrappers on top of c_FlushFileBuffers, and
c_CloseHandle. These wrappers are:

  closeHandleCheckInvalidHandle :: HANDLE -> IO Bool
  flushFileBuffersCheckInvalidHandle :: HANDLE -> IO Bool

Both of these wrappers return False when successful and True when handle
was invalid (ERROR_INVALID_HANDLE). We hadn't had to reimplement
disconnectNamedPipe, since flushFileBuffers is always called before
disconnectNamedPipe, therefore, we can shortcircuit before calling
disconnectNamedPipe.

Functions closePipe and disconnectPipe now have different return value:

  closePipe :: PipeHandle -> IO Bool
  disconnectPipe :: PipeHandle -> IO Bool

Meaning of the Bool return value is the same as with
closeHandleCheckInvalidHandle, and flushFileBuffersCheckInvalidHandle.

Server and client functions were updated to handle the new return value of
the functions mentioned above.