Fix "file/pipe not exists" errors
There was a race condition in runPipeServer that caused all pipe
instances to vanish for a short amount of time confusing clients with
"file/pipe not exists" errors (even WaitNamedPipe would return
immediately with an error).

This also fixes serverAfterBindPipe being invoked multiple times, which
was quite confusing as it differed from how runTCPServer, runUnixServer
behave.